Using i32 instead of u32 for world coordinates

+++ core/src/utils.rs
@@ -1,8 +1,8 @@
 use crate::Point3;
 
-pub fn is_near(a: &Point3<u32>, b: &Point3<u32>) -> bool {
-    let dist_x = (a.x as i32 - b.x as i32).abs() as u32;
-    let dist_y = (a.y as i32 - b.y as i32).abs() as u32;
+pub fn is_near(a: &Point3<i32>, b: &Point3<i32>) -> bool {
+    let dist_x = (a.x - b.x).abs();
+    let dist_y = (a.y - b.y).abs();
 
     dist_x + dist_y <= 1
 }
